Force refresh CSS for clients
I've messed up, so currently the main website(production) is showing
development. When users go to visit the OS site, the CSS will not
refresh and will make the site look buggy. Adding parameters to the end
of the CSS url should make the cache re-download the CSS file. Hopefully
this is a temporary thing until production can be reversed.