Click on this commit!
Nice! Try this out:

On your mobile device (iPhone or Android), go to https://www.opensprites.org and tap add to home screen. This code makes it so that when you click the icon on your home screen, it will open an app version of OpenSprites. (Also part of it just changes the title bar color). :P