Temporary missing file extention bug fix
The name of the asset was not being prefixed before the file extention causing files to be unrecognisable when downloaded. This temporary fix removes the asset name altogether, replacing it with a static name. This file should be updated futher down the line so that asset names are in the download name. The ability to download GIFs has also been implemented.